+++ tests/from_json_test.py
@@ -11,7 +11,7 @@ def test_from_json(ds_local):
     with open(tmp, 'w') as f:
         f.write(pandas_df.to_json())
 
-    tmp_df = vaex.from_json(tmp, copy_index=False)
+    tmp_df = vaex.from_json(tmp)
 
     assert set(tmp_df.column_names) == set(df.column_names)
     assert len(tmp_df) == len(df)