+++ CHANGELOG.md
@@ -9,6 +9,7 @@
      * Default value of `virtual` is True in method `df.export`, `df.to_dict`, `df.to_items`, `df.to_arrays`.
      * df.dtype is a property, to get data types for expressions, use df.data_type(), df.expr.dtype is still behaving the same
      * df.categorize takes min_value and max_value, and no longer needs the check argument, also the labels do not have to be strings.
+     * vaex.open/from_csv etc does not copy the pandas index by default [#756](https://github.com/vaexio/vaex/pull/756)
 
 # vaex-core 2.0.0-dev
    * Performance